// multi/src/app/model/view.ts
/**********************************************************************************
* File Name   :   View.ts
* Description :   View
*
* Copyright (c) MultiSight™ © 2016 Schneider Electric 
* --------------------------------------------------------------------------------
* Date             Author              Comment
* --------------------------------------------------------------------------------
* 1-Nov-2016      QuEST Team          Initial version created
**********************************************************************************/

export class View {
    id: string;
    name: string;
    shortName: string;
    dtnSiteID: string;
    createdTime: string;
    lastEquipmentChangeTime: string;
    timeZone: string;
    lastSeenTime: string;
    gatewayId: string;
    type: string;
    cameraList: any;

}